requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
psycopg2-binary>=2.9.9
Pillow>=10.1.0
python-decouple>=3.8
gunicorn>=21.2.0
whitenoise>=6.6.0
pyTelegramBotAPI>=4.14.0
dj-database-url>=2.1.0

